+++ bench/speed/Main.hs
@@ -15,6 +15,7 @@ main = defaultMain
   , bparser "data/bench-ordered-list.md"
   , bparser "data/bench-blockquote.md"
   , bparser "data/bench-paragraph.md"
+  , bparser "data/table.md"
   , bparser "data/comprehensive.md"
   ]