+++ CHANGELOG.md
@@ -10,8 +10,7 @@
 * Fixed a couple of bugs in the parser for reference definitions.
 
 * Now link and image titles may contain newline character as per the Common
-  Mark spec. They also may contain blank lines, which is against the spec
-  though.
+  Mark spec.
 
 ## MMark 0.0.3.0